perf: add MILK_RESTRICT, MILK_ASSUME_ALIGNED, and fix uint32 overflow in coremods (#273)
- imfunctions.c: Add MILK_RESTRICT + MILK_ASSUME_ALIGNED + const to all
  4 fast-path pointer aliases (float/double same-type and mixed-type).
  This is the central dispatch for all image arithmetic.

- image_stats.c: Replace bare 'restrict' with MILK_RESTRICT +
  MILK_ASSUME_ALIGNED in 4 signed-integer branches of
  arith_image_max_IMGID, matching the existing unsigned-type pattern.

- image_copy.c: Replace memcpy with __builtin_memcpy in 2 bulk pixel
  copy paths (copy_image_ID_IMGID, COREMOD_MEMORY_cp2shm_IMGID).

- stream_diff.c: Fix uint32_t overflow in xysize computation; add
  MILK_IVDEP to both element-wise diff loops.

- stream_halfimdiff.c: Fix same uint32_t overflow in xysize.

- image_pixremap.c: Remove incorrect MILK_IVDEP from 2 counting loops
  (nbpix++ is a loop-carried dependency); standardize __restrict to
  MILK_RESTRICT; fix uint32_t overflow in loop bounds.

- image_cropmask.c: Add MILK_IVDEP to inner pixel multiply loop.

+++ src/coremods/COREMOD_memory/stream_halfimdiff.c
@@ -82,7 +82,7 @@ imageID MILK_HOT COREMOD_MEMORY_stream_halfimDiff(
 
     uint32_t xsize  = xsizein;
     uint32_t ysize  = ysizein / 2;
-    uint64_t xysize = xsize * ysize;
+    uint64_t xysize = (uint64_t)xsize * ysize;
 
     uint8_t datatype    = img0.md->datatype;
     uint8_t datatypeout = _DATATYPE_FLOAT;
